*Minor change to satisfy MinGW. *Improved the leading silence re-encode feature so that it will prompt for the target bitrate to encode with.  If this prompt is cancelled, EOF will automatically fall back to the previous behavior of automatically basing the re-encode bitrate as the current chart audio file's bitrate. *Changed GP import so that if slide in from above/below notes are encountered, they are highlighted and a warning is given that Rocksmith doesn't support this technique, so the author can choose to edit EOF's default interpretation (unpitched slide of one fret) for such notes if desired.

+++ src/menu/song.c
@@ -4124,7 +4124,7 @@ int eof_menu_song_export_song_preview(void)
 			#ifdef ALLEGRO_WINDOWS
 				(void) uszprintf(syscommand, (int) sizeof(syscommand), "oggenc2 --quiet -q %s --resample 44100 -s 0 \"%s%s\" -o \"%spreview.ogg\"", eof_ogg_quality[(int)eof_ogg_setting], targetpath, wavname, targetpath);
 			#else
-				(void) uszprintf(syscommand, (int) sizeof(syscommand), "oggenc --quiet -q %s --resample 44100 -s 0 \"%s%s\" -o \"%spreview.ogg\"", eof_ogg_quality[(int)eof_ogg_setting], targetpath, wavname, targetpath);
+				(void) uszprintf(syscommand, (int) sizeof(syscommand), "oggenc --quiet -q %s --resample 44100 -s 0 \"%s%s\" -o \"%spreview.ogg\"", eof_ogg_quality[eof_ogg_setting], targetpath, wavname, targetpath);
 			#endif
 			(void) eof_system(syscommand);
 			if(!exists(targetpath))
